+++ src/graph.h
@@ -5,15 +5,15 @@
 
 class graph {
 private:
-    Vec2f origin;
-    std::vector<float> data;
-    Vec2f dimens{500, 200};//500px wide, 200px tall
-    Vec2f axis_scale{100, 100};
+    Vec2d origin;
+    std::vector<double> data;
+    Vec2d dimens{500, 200};//500px wide, 200px tall
+    Vec2d axis_scale{100, 100};
     void rescale();
     void resize();
 public:
-    graph(Vec2f o) : origin(o) {}
-    void add_data(float a);
+    graph(Vec2d o) : origin(o) {}
+    void add_data(double a);
     void draw();
 };
